+++ samples/android-simple/src/main/java/io/github/tapcard/sample/android/simple/MainActivity.java
@@ -3,7 +3,7 @@ package io.github.tapcard.sample.android.simple;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.text.format.DateFormat;
 import android.widget.TextView;