+++ samples/android/src/main/java/io/github/tapcard/sample/android/adapter/ViewPagerAdapter.java
@@ -2,8 +2,8 @@ package io.github.tapcard.sample.android.adapter;
 
 import java.util.List;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
 import io.github.tapcard.sample.android.fragment.viewPager.IFragment;