+++ samples/android/src/main/java/io/github/tapcard/sample/android/preference/PreferenceFragment.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package android.support.v4.preference;
+package io.github.tapcard.sample.android.preference;
 
 import io.github.tapcard.sample.android.R;
 
@@ -26,7 +26,7 @@ import android.preference.Preference;
 import android.preference.PreferenceGroup;
 import android.preference.PreferenceManager;
 import android.preference.PreferenceScreen;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;