+++ samples/android/src/main/java/io/github/tapcard/sample/android/utils/NFCUtils.java
@@ -84,8 +84,13 @@ public class NFCUtils {
 	public NFCUtils(final Activity pActivity) {
 		mActivity = pActivity;
 		mNfcAdapter = NfcAdapter.getDefaultAdapter(mActivity);
+		int pendingIntentFlags = PendingIntent.FLAG_UPDATE_CURRENT;
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.S) {
+			pendingIntentFlags |= PendingIntent.FLAG_MUTABLE;
+		}
 		mPendingIntent = PendingIntent.getActivity(mActivity, 0,
-				new Intent(mActivity, mActivity.getClass()).addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP), 0);
+				new Intent(mActivity, mActivity.getClass()).addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP),
+				pendingIntentFlags);
 	}
 
 	/**