+++ samples/android/src/main/java/io/github/tapcard/sample/android/view/SlidingTabLayout.java
@@ -20,8 +20,8 @@ import android.content.Context;
 import android.graphics.Color;
 import android.graphics.Typeface;
 import android.os.Build;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 import android.util.TypedValue;
 import android.view.Gravity;
@@ -37,7 +37,7 @@ import io.github.tapcard.sample.android.adapter.ViewPagerAdapter;
  * To be used with ViewPager to provide a tab indicator component which give constant feedback as to the user's scroll progress.
  * <p>
  * To use the component, simply add it to your view hierarchy. Then in your {@link android.app.Activity} or
- * {@link android.support.v4.app.Fragment} call {@link #setViewPager(ViewPager)} providing it the ViewPager this layout is being
+ * {@link androidx.fragment.app.Fragment} call {@link #setViewPager(ViewPager)} providing it the ViewPager this layout is being
  * used for.
  * <p>
  * The colors can be customized in two ways. The first and simplest is to provide an array of colors via